fix(Thema): Use a specific version of Thema as a dependency.
In /open the HTML in the iframe is generated by Encoda, including the `<head>`. Therefore the `THEMA_VERSION` is not applied until `setIframeTheme` is called (i.e. when the theme is changed). I propose NOT to address this now, but instead wait for a planned refactoring of /open. Also, Encoda will be updated to Thema@2 very soon anyway.

Closes #367.