fix(Form fields): Use the field name as the id
Using a randomly generated id for fields broke HTMX events with in-browser errors like this:

Uncaught DOMException: Failed to execute 'querySelector' on 'Element': 'INPUT[id=94f4ea96-0821-4380-abf3-584959842d9e]' is not a valid selector.

This for example broke the error messages on forms when there was an error in the field.